Modify module tests
