+++ sdk-core/src/test/java/care/data4life/sdk/RecordServiceFetchRecordsModuleTest.kt
@@ -30,7 +30,6 @@ import care.data4life.sdk.model.Record
 import care.data4life.sdk.network.NetworkingContract
 import care.data4life.sdk.network.model.EncryptedKey
 import care.data4life.sdk.network.model.EncryptedRecord
-import care.data4life.sdk.network.util.SearchTagsBuilder
 import care.data4life.sdk.record.RecordContract
 import care.data4life.sdk.tag.Annotations
 import care.data4life.sdk.tag.TagCryptoService
@@ -43,12 +42,8 @@ import care.data4life.sdk.test.util.GenericTestDataProvider.ARBITRARY_DATA_KEY
 import care.data4life.sdk.test.util.GenericTestDataProvider.CLIENT_ID
 import care.data4life.sdk.test.util.GenericTestDataProvider.COMMON_KEY_ID
 import care.data4life.sdk.test.util.GenericTestDataProvider.CREATION_DATE
-import care.data4life.sdk.test.util.GenericTestDataProvider.OFFSET
-import care.data4life.sdk.test.util.GenericTestDataProvider.PAGE_SIZE
 import care.data4life.sdk.test.util.GenericTestDataProvider.PARTNER_ID
 import care.data4life.sdk.test.util.GenericTestDataProvider.RECORD_ID
-import care.data4life.sdk.test.util.GenericTestDataProvider.RECORD_ID_LEGACY_JS
-import care.data4life.sdk.test.util.GenericTestDataProvider.RECORD_ID_LEGACY_KMP
 import care.data4life.sdk.test.util.GenericTestDataProvider.UPDATE_DATE
 import care.data4life.sdk.test.util.GenericTestDataProvider.USER_ID
 import care.data4life.sdk.test.util.TestResourceHelper
@@ -61,7 +56,6 @@ import io.reactivex.Observable
 import io.reactivex.Single
 import org.junit.Before
 import org.junit.Test
-import org.threeten.bp.LocalDate
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import care.data4life.fhir.r4.model.DocumentReference as Fhir4DocumentReference
@@ -328,181 +322,6 @@ class RecordServiceFetchRecordsModuleTest {
                 )
             }
         }
-    }
-
-    private fun runFhirBatchFlow(
-        serializedResources: Triple<String, String, String>,
-        tags: Tags,
-        annotations: Annotations = emptyList(),
-        tagEncryptionKeyCalls: Int = 4,
-        useStoredCommonKey: Boolean = true,
-        commonKey: Pair<String, GCKey> = COMMON_KEY_ID to this.commonKey,
-        dataKey: Pair<GCKey, EncryptedKey> = this.dataKey to encryptedDataKey,
-        attachmentKey: Pair<GCKey, EncryptedKey>? = this.attachmentKey to encryptedAttachmentKey,
-        tagEncryptionKey: GCKey = this.tagEncryptionKey,
-        recordIds: Triple<String, String, String> = Triple(RECORD_ID, RECORD_ID_LEGACY_KMP, RECORD_ID_LEGACY_JS),
-        userId: String = USER_ID,
-        alias: String = ALIAS,
-        creationDate: String = CREATION_DATE,
-        updateDate: String = UPDATE_DATE,
-        startDate: String? = null,
-        endDate: String? = null,
-        pageSize: Int = PAGE_SIZE,
-        offset: Int = OFFSET
-    ) {
-        val (encodedTags, legacyKMPTags, legacyJSTags) = flowHelper.prepareCompatibilityTags(tags)
-        val (encodedAnnotations, legacyKMPAnnotations, legacyJSAnnotations) = flowHelper.prepareCompatibilityAnnotations(
-            annotations
-        )
-
-        val searchTags = SearchTagsBuilder.newBuilder()
-            .let { flowHelper.buildExpectedTagGroups(it, encodedTags, legacyKMPTags, legacyJSTags) }
-            .let { flowHelper.buildExpectedTagGroups(it, encodedAnnotations, legacyKMPAnnotations, legacyJSAnnotations) }
-            .seal()
-            .tagGroups
-
-        val encryptedRecord = flowHelper.prepareEncryptedRecord(
-            recordIds.first,
-            serializedResources.first,
-            encodedTags,
-            encodedAnnotations,
-            commonKey.first,
-            dataKey.second,
-            attachmentKey?.second,
-            creationDate,
-            updateDate
-        )
-
-        val encryptedKMPLegacyRecord = flowHelper.prepareEncryptedRecord(
-            recordIds.second,
-            serializedResources.second,
-            legacyKMPTags,
-            legacyKMPAnnotations,
-            commonKey.first,
-            dataKey.second,
-            attachmentKey?.second,
-            creationDate,
-            updateDate
-        )
-
-        val encryptedJSLegacyRecord = flowHelper.prepareEncryptedRecord(
-            recordIds.third,
-            serializedResources.third,
-            legacyJSTags,
-            legacyJSAnnotations,
-            commonKey.first,
-            dataKey.second,
-            attachmentKey?.second,
-            creationDate,
-            updateDate
-        )
-
-        runBatchFlow(
-            serializedResources,
-            listOf(encryptedRecord, encryptedKMPLegacyRecord, encryptedJSLegacyRecord),
-            searchTags,
-            flowHelper.mergeTags(encodedTags, legacyKMPTags, legacyJSTags),
-            flowHelper.mergeTags(encodedAnnotations, legacyKMPAnnotations, legacyJSAnnotations),
-            tagEncryptionKeyCalls,
-            useStoredCommonKey,
-            commonKey,
-            dataKey,
-            attachmentKey,
-            tagEncryptionKey,
-            userId,
-            alias,
-            startDate,
-            endDate,
-            pageSize,
-            offset
-        )
-    }
-
-    private fun runDataBatchFlow(
-        serializedResources: Triple<String, String, String>,
-        tags: Tags,
-        annotations: Annotations = emptyList(),
-        tagEncryptionKeyCalls: Int = 4,
-        useStoredCommonKey: Boolean = true,
-        commonKey: Pair<String, GCKey> = COMMON_KEY_ID to this.commonKey,
-        dataKey: Pair<GCKey, EncryptedKey> = this.dataKey to encryptedDataKey,
-        attachmentKey: Pair<GCKey, EncryptedKey>? = this.attachmentKey to encryptedAttachmentKey,
-        tagEncryptionKey: GCKey = this.tagEncryptionKey,
-        recordIds: Triple<String, String, String> = Triple(RECORD_ID, RECORD_ID_LEGACY_KMP, RECORD_ID_LEGACY_JS),
-        userId: String = USER_ID,
-        alias: String = ALIAS,
-        creationDate: String = CREATION_DATE,
-        updateDate: String = UPDATE_DATE,
-        startDate: String? = null,
-        endDate: String? = null,
-        pageSize: Int = PAGE_SIZE,
-        offset: Int = OFFSET
-    ) {
-        val (encodedTags, legacyKMPTags, legacyJSTags) = flowHelper.prepareCompatibilityTags(tags)
-        val (encodedAnnotations, legacyKMPAnnotations, legacyJSAnnotations) = flowHelper.prepareCompatibilityAnnotations(
-            annotations
-        )
-        val searchTags = SearchTagsBuilder.newBuilder()
-            .let { flowHelper.buildExpectedTagGroups(it, encodedTags, legacyKMPTags, legacyJSTags) }
-            .let { flowHelper.buildExpectedTagGroups(it, encodedAnnotations, legacyKMPAnnotations, legacyJSAnnotations) }
-            .seal()
-            .tagGroups
-
-        val encryptedRecord = flowHelper.prepareEncryptedRecord(
-            recordIds.first,
-            serializedResources.first,
-            encodedTags,
-            encodedAnnotations,
-            commonKey.first,
-            dataKey.second,
-            attachmentKey?.second,
-            creationDate,
-            updateDate
-        )
-
-        val encryptedKMPLegacyRecord = flowHelper.prepareEncryptedRecord(
-            recordIds.second,
-            serializedResources.second,
-            legacyKMPTags,
-            legacyKMPAnnotations,
-            commonKey.first,
-            dataKey.second,
-            attachmentKey?.second,
-            creationDate,
-            updateDate
-        )
-
-        val encryptedJSLegacyRecord = flowHelper.prepareEncryptedRecord(
-            recordIds.third,
-            serializedResources.third,
-            legacyJSTags,
-            legacyJSAnnotations,
-            commonKey.first,
-            dataKey.second,
-            attachmentKey?.second,
-            creationDate,
-            updateDate
-        )
-
-        runBatchFlow(
-            serializedResources,
-            listOf(encryptedRecord, encryptedKMPLegacyRecord, encryptedJSLegacyRecord),
-            searchTags,
-            flowHelper.mergeTags(encodedTags, legacyKMPTags, legacyJSTags),
-            flowHelper.mergeTags(encodedAnnotations, legacyKMPAnnotations, legacyJSAnnotations),
-            tagEncryptionKeyCalls,
-            useStoredCommonKey,
-            commonKey,
-            dataKey,
-            attachmentKey,
-            tagEncryptionKey,
-            userId,
-            alias,
-            startDate,
-            endDate,
-            pageSize,
-            offset
-        )
     }
 
     // Fetch
@@ -807,854 +626,4 @@ class RecordServiceFetchRecordsModuleTest {
             actual = String(result.resource.value)
         )
     }
-
-    // search
-    // FHIR 3
-    @Test
-    fun `Given, fetchFhir3Records is called, with its appropriate payloads, it returns a List of Records`() {
-        // Given
-        val resourceType = "DocumentReference"
-        val tags = mapOf(
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID,
-            "fhirversion" to "3.0.1",
-            "resourcetype" to resourceType
-        )
-
-        val template = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID,
-            PARTNER_ID
-        )
-
-        val template2 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_KMP,
-            PARTNER_ID
-        )
-
-        val template3 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_JS,
-            PARTNER_ID
-        )
-
-        val resource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template
-        ) as Fhir3DocumentReference
-
-        val legacyKMPResource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template2
-        ) as Fhir3DocumentReference
-
-        val legacyJSResource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template3
-        ) as Fhir3DocumentReference
-
-        legacyKMPResource.description = "legacyKMPRecord"
-        legacyJSResource.description = "legacyKMPRecord"
-
-        runFhirBatchFlow(
-            serializedResources = Triple(
-                SdkFhirParser.fromResource(resource),
-                SdkFhirParser.fromResource(legacyKMPResource),
-                SdkFhirParser.fromResource(legacyJSResource)
-            ),
-            tags = tags,
-            useStoredCommonKey = false
-        )
-
-        // When
-        val result = recordService.searchFhir3Records(
-            USER_ID,
-            Fhir3DocumentReference::class.java,
-            emptyList(),
-            null,
-            null,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[0].resource),
-            expected = SdkFhirParser.fromResource(resource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[1].resource),
-            expected = SdkFhirParser.fromResource(legacyKMPResource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[2].resource),
-            expected = SdkFhirParser.fromResource(legacyJSResource)
-        )
-        assertTrue(result[0].annotations.isNullOrEmpty())
-        assertTrue(result[1].annotations.isNullOrEmpty())
-        assertTrue(result[2].annotations.isNullOrEmpty())
-    }
-
-    @Test
-    fun `Given, fetchFhir3Records is called, with its appropriate payloads, it returns a List of Records filtered by Annotations`() {
-        // Given
-        val resourceType = "DocumentReference"
-        val tags = mapOf(
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID,
-            "fhirversion" to "3.0.1",
-            "resourcetype" to resourceType
-        )
-
-        val annotations = listOf(
-            "wow",
-            "it",
-            "works",
-            "and",
-            "like_a_duracell_häsi"
-        )
-
-        val template = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID,
-            PARTNER_ID
-        )
-
-        val template2 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_KMP,
-            PARTNER_ID
-        )
-
-        val template3 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_JS,
-            PARTNER_ID
-        )
-
-        val resource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template
-        ) as Fhir3DocumentReference
-
-        val legacyKMPResource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template2
-        ) as Fhir3DocumentReference
-
-        val legacyJSResource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template3
-        ) as Fhir3DocumentReference
-
-        legacyKMPResource.description = "legacyRecord"
-        legacyJSResource.description = "legacyRecord"
-
-        runFhirBatchFlow(
-            serializedResources = Triple(
-                SdkFhirParser.fromResource(resource),
-                SdkFhirParser.fromResource(legacyKMPResource),
-                SdkFhirParser.fromResource(legacyJSResource)
-            ),
-            tags = tags,
-            annotations = annotations
-        )
-
-        // When
-        val result = recordService.searchFhir3Records(
-            USER_ID,
-            Fhir3DocumentReference::class.java,
-            annotations,
-            null,
-            null,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[0].resource),
-            expected = SdkFhirParser.fromResource(resource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[1].resource),
-            expected = SdkFhirParser.fromResource(legacyKMPResource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[2].resource),
-            expected = SdkFhirParser.fromResource(legacyJSResource)
-        )
-        assertEquals(
-            actual = result[0].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[1].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[2].annotations,
-            expected = annotations
-        )
-    }
-
-    /*
-    @Test
-    fun `Given, fetchFhir3Records is called, with its appropriate payloads, it returns a List of Records filtered by the provided date parameter and Annotations`() {
-        // Given
-        val resourceType = "DocumentReference"
-        val tags = mapOf(
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID,
-            "fhirversion" to "3.0.1",
-            "resourcetype" to resourceType
-        )
-
-        val annotations = listOf(
-            "wow",
-            "it",
-            "works",
-            "and",
-            "like_a_duracell_häsi"
-        )
-
-        val startDate = LocalDate.of(2020, 1, 1)
-        val endDate = LocalDate.of(2020, 12, 1)
-
-        val template = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID,
-            PARTNER_ID
-        )
-
-        val template2 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_KMP,
-            PARTNER_ID
-        )
-
-        val template3 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_JS,
-            PARTNER_ID
-        )
-
-        val resource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template
-        ) as Fhir3DocumentReference
-
-        val legacyKMPResource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template2
-        ) as Fhir3DocumentReference
-
-        val legacyJSResource = SdkFhirParser.toFhir<Fhir3Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_3.version,
-            template3
-        ) as Fhir3DocumentReference
-
-        legacyKMPResource.description = "legacyRecord"
-        legacyJSResource.description = "legacyRecord"
-
-        runFhirBatchFlow(
-            serializedResources = Triple(
-                SdkFhirParser.fromResource(resource),
-                SdkFhirParser.fromResource(legacyKMPResource),
-                SdkFhirParser.fromResource(legacyJSResource),
-            ),
-            tags = tags,
-            annotations = annotations,
-            startDate = startDate.toString(),
-            endDate = endDate.toString()
-        )
-
-        // When
-        val result = recordService.searchFhir3Records(
-            USER_ID,
-            Fhir3DocumentReference::class.java,
-            annotations,
-            startDate,
-            endDate,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[0].resource),
-            expected = SdkFhirParser.fromResource(resource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[1].resource),
-            expected = SdkFhirParser.fromResource(legacyKMPResource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[2].resource),
-            expected = SdkFhirParser.fromResource(legacyJSResource)
-        )
-        assertEquals(
-            actual = result[0].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[1].annotations,
-            expected = annotations
-        )
-
-        assertEquals(
-            actual = result[2].annotations,
-            expected = annotations
-        )
-    }
-
-    // FHIR 4
-    @Test
-    fun `Given, fetchFhir4Records is called, with its appropriate payloads, it returns a List of Records`() {
-        // Given
-        val resourceType = "DocumentReference"
-        val tags = mapOf(
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID,
-            "fhirversion" to "4.0.1",
-            "resourcetype" to resourceType
-        )
-
-        val template = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID,
-            PARTNER_ID
-        )
-
-        val template2 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_KMP,
-            PARTNER_ID
-        )
-
-        val template3 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_JS,
-            PARTNER_ID
-        )
-
-        val resource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template
-        ) as Fhir4DocumentReference
-
-        val legacyKMPResource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template2
-        ) as Fhir4DocumentReference
-
-        val legacyJSResource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template3
-        ) as Fhir4DocumentReference
-
-        legacyKMPResource.description = "legacyKMPRecord"
-        legacyJSResource.description = "legacyJSRecord"
-
-        runFhirBatchFlow(
-            serializedResources = Triple(
-                SdkFhirParser.fromResource(resource),
-                SdkFhirParser.fromResource(legacyKMPResource),
-                SdkFhirParser.fromResource(legacyJSResource)
-            ),
-            tags = tags,
-            useStoredCommonKey = false
-        )
-
-        // When
-        val result = recordService.searchFhir4Records(
-            USER_ID,
-            Fhir4DocumentReference::class.java,
-            emptyList(),
-            null,
-            null,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[0].resource),
-            expected = SdkFhirParser.fromResource(resource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[1].resource),
-            expected = SdkFhirParser.fromResource(legacyKMPResource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[2].resource),
-            expected = SdkFhirParser.fromResource(legacyJSResource)
-        )
-        assertTrue(result[0].annotations.isNullOrEmpty())
-        assertTrue(result[1].annotations.isNullOrEmpty())
-        assertTrue(result[2].annotations.isNullOrEmpty())
-    }
-
-    @Test
-    fun `Given, fetchFhir4Records is called, with its appropriate payloads, it returns a List of Records filtered by Annotations`() {
-        // Given
-        val resourceType = "DocumentReference"
-        val tags = mapOf(
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID,
-            "fhirversion" to "4.0.1",
-            "resourcetype" to resourceType
-        )
-
-        val annotations = listOf(
-            "wow",
-            "it",
-            "works",
-            "and",
-            "like_a_duracell_häsi"
-        )
-
-        val template = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID,
-            PARTNER_ID
-        )
-
-        val template2 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_KMP,
-            PARTNER_ID
-        )
-
-        val template3 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_JS,
-            PARTNER_ID
-        )
-
-        val resource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template
-        ) as Fhir4DocumentReference
-
-        val legacyKMPResource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template2
-        ) as Fhir4DocumentReference
-
-        val legacyJSResource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template3
-        ) as Fhir4DocumentReference
-
-        legacyKMPResource.description = "legacyKMPRecord"
-        legacyJSResource.description = "legacyJSRecord"
-
-        runFhirBatchFlow(
-            serializedResources = Triple(
-                SdkFhirParser.fromResource(resource),
-                SdkFhirParser.fromResource(legacyKMPResource),
-                SdkFhirParser.fromResource(legacyJSResource)
-            ),
-            tags = tags,
-            annotations = annotations
-        )
-
-        // When
-        val result = recordService.searchFhir4Records(
-            USER_ID,
-            Fhir4DocumentReference::class.java,
-            annotations,
-            null,
-            null,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[0].resource),
-            expected = SdkFhirParser.fromResource(resource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[1].resource),
-            expected = SdkFhirParser.fromResource(legacyKMPResource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[2].resource),
-            expected = SdkFhirParser.fromResource(legacyJSResource)
-        )
-        assertEquals(
-            actual = result[0].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[1].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[2].annotations,
-            expected = annotations
-        )
-    }
-
-    @Test
-    fun `Given, fetchFhir4Records is called, with its appropriate payloads, it returns a List of Records filtered by the provided date parameter and Annotations`() {
-        // Given
-        val resourceType = "DocumentReference"
-        val tags = mapOf(
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID,
-            "fhirversion" to "4.0.1",
-            "resourcetype" to resourceType
-        )
-
-        val annotations = listOf(
-            "wow",
-            "it",
-            "works",
-            "and",
-            "like_a_duracell_häsi"
-        )
-
-        val startDate = LocalDate.of(2020, 1, 1)
-        val endDate = LocalDate.of(2020, 12, 1)
-
-        val template = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID,
-            PARTNER_ID
-        )
-
-        val template2 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_KMP,
-            PARTNER_ID
-        )
-
-        val template3 = TestResourceHelper.loadTemplate(
-            "common",
-            "documentReference-without-attachment-template",
-            RECORD_ID_LEGACY_JS,
-            PARTNER_ID
-        )
-
-        val resource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template
-        ) as Fhir4DocumentReference
-
-        val legacyKMPResource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template2
-        ) as Fhir4DocumentReference
-
-        val legacyJSResource = SdkFhirParser.toFhir<Fhir4Resource>(
-            resourceType,
-            FhirContract.FhirVersion.FHIR_4.version,
-            template3
-        ) as Fhir4DocumentReference
-
-        legacyKMPResource.description = "legacyKMPRecord"
-        legacyJSResource.description = "legacyJSRecord"
-
-        runFhirBatchFlow(
-            serializedResources = Triple(
-                SdkFhirParser.fromResource(resource),
-                SdkFhirParser.fromResource(legacyKMPResource),
-                SdkFhirParser.fromResource(legacyJSResource),
-            ),
-            tags = tags,
-            annotations = annotations,
-            startDate = startDate.toString(),
-            endDate = endDate.toString()
-        )
-
-        // When
-        val result = recordService.searchFhir4Records(
-            USER_ID,
-            Fhir4DocumentReference::class.java,
-            annotations,
-            startDate,
-            endDate,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[0].resource),
-            expected = SdkFhirParser.fromResource(resource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[1].resource),
-            expected = SdkFhirParser.fromResource(legacyKMPResource)
-        )
-        assertEquals(
-            actual = SdkFhirParser.fromResource(result[2].resource),
-            expected = SdkFhirParser.fromResource(legacyJSResource)
-        )
-        assertEquals(
-            actual = result[0].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[1].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[2].annotations,
-            expected = annotations
-        )
-    }
-
-    // Arbitrary Data
-    @Test
-    fun `Given, fetchDataRecords is called, with its appropriate payloads, it returns a List of Records`() {
-        // Given
-        val resource = "The never ending story."
-        val tags = mapOf(
-            "flag" to ARBITRARY_DATA_KEY,
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID
-        )
-
-        runDataBatchFlow(
-            serializedResources = Triple(resource, resource, resource),
-            tags = tags,
-            useStoredCommonKey = false
-        )
-
-        // When
-        val result = recordService.searchDataRecords(
-            USER_ID,
-            emptyList(),
-            null,
-            null,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = String(result[0].resource.value),
-            expected = resource
-        )
-        assertEquals(
-            actual = String(result[1].resource.value),
-            expected = resource
-        )
-        assertEquals(
-            actual = String(result[2].resource.value),
-            expected = resource
-        )
-
-        assertTrue(result[0].annotations.isNullOrEmpty())
-        assertTrue(result[1].annotations.isNullOrEmpty())
-        assertTrue(result[2].annotations.isNullOrEmpty())
-    }
-
-    @Test
-    fun `Given, fetchDataRecords is called, with its appropriate payloads, it returns a List of Records filtered by Annotations`() {
-        // Given
-        val resource = "The never ending story."
-        val legacyKMPResource = "Completely new design, ey."
-        val legacyJSResource = "And another bug, ey."
-
-        val tags = mapOf(
-            "flag" to ARBITRARY_DATA_KEY,
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID
-        )
-
-        val annotations = listOf(
-            "wow",
-            "it",
-            "works",
-            "and",
-            "like_a_duracell_häsi"
-        )
-
-        runDataBatchFlow(
-            serializedResources = Triple(resource, legacyKMPResource, legacyJSResource),
-            tags = tags,
-            annotations = annotations
-        )
-
-        // When
-        val result = recordService.searchDataRecords(
-            USER_ID,
-            annotations,
-            null,
-            null,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = String(result[0].resource.value),
-            expected = resource
-        )
-        assertEquals(
-            actual = String(result[1].resource.value),
-            expected = legacyKMPResource
-        )
-        assertEquals(
-            actual = String(result[2].resource.value),
-            expected = legacyJSResource
-        )
-        assertEquals(
-            actual = result[0].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[1].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[2].annotations,
-            expected = annotations
-        )
-    }
-
-    @Test
-    fun `Given, fetchDataRecords is called, with its appropriate payloads, it returns a List of Records filtered by the provided date parameter and Annotations`() {
-        // Given
-        val resource = "The never ending story."
-        val legacyKMPResource = "Completely new design, ey."
-        val legacyJSResource = "And another bug, ey."
-
-        val tags = mapOf(
-            "flag" to ARBITRARY_DATA_KEY,
-            "partner" to PARTNER_ID,
-            "client" to CLIENT_ID
-        )
-
-        val annotations = listOf(
-            "wow",
-            "it",
-            "works",
-            "and",
-            "like_a_duracell_häsi"
-        )
-
-        val startDate = LocalDate.of(2020, 1, 1)
-        val endDate = LocalDate.of(2020, 12, 1)
-
-        runDataBatchFlow(
-            serializedResources = Triple(resource, legacyKMPResource, legacyJSResource),
-            tags = tags,
-            annotations = annotations,
-            startDate = startDate.toString(),
-            endDate = endDate.toString()
-        )
-
-        // When
-        val result = recordService.searchDataRecords(
-            USER_ID,
-            annotations,
-            startDate,
-            endDate,
-            PAGE_SIZE,
-            OFFSET
-        ).blockingGet()
-
-        // Then
-        assertEquals(
-            expected = 3,
-            actual = result.size
-        )
-        assertEquals(
-            actual = String(result[0].resource.value),
-            expected = resource
-        )
-        assertEquals(
-            actual = String(result[1].resource.value),
-            expected = legacyKMPResource
-        )
-        assertEquals(
-            actual = String(result[2].resource.value),
-            expected = legacyJSResource
-        )
-        assertEquals(
-            actual = result[0].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[1].annotations,
-            expected = annotations
-        )
-        assertEquals(
-            actual = result[2].annotations,
-            expected = annotations
-        )
-    }*/
 }
