+++ sdk-core/src/test/java/care/data4life/sdk/model/RecordMapperTest.kt
@@ -20,6 +20,7 @@ import care.data4life.crypto.GCKey
 import care.data4life.sdk.call.DataRecord
 import care.data4life.sdk.call.Fhir4Record
 import care.data4life.sdk.data.DataResource
+import care.data4life.sdk.date.SdkDateTimeFormatter
 import care.data4life.sdk.fhir.Fhir3Resource
 import care.data4life.sdk.fhir.Fhir4Resource
 import care.data4life.sdk.lang.CoreRuntimeException
@@ -31,7 +32,6 @@ import care.data4life.sdk.network.model.DecryptedRecord
 import care.data4life.sdk.network.model.NetworkModelContract.DecryptedBaseRecord
 import care.data4life.sdk.tag.Annotations
 import care.data4life.sdk.tag.Tags
-import care.data4life.sdk.date.SdkDateTimeFormatter
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkObject
